split up blog post into components, minor fixes
