+++ app/blog/[slug]/TextButton.js
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+const TextButton = styled.a`
+  cursor: pointer;
+`;
+
+export default TextButton;
