+++ app/blog/[slug]/page.module.css
@@ -24,6 +24,7 @@
 
 .buttons {
   display: flex;
+  flex-wrap: wrap;
   gap: 1rem;
 }
