+++ package.json
@@ -17,7 +17,7 @@
     "emoji-picker-react": "^4.4.8",
     "marked-react": "^1.3.2",
     "next": "13.2.4",
-    "react": "18.2.0",
+    "react": "^18.2.0",
     "react-dom": "18.2.0",
     "react-icons": "^4.8.0",
     "styled-components": "^5.3.9"
@@ -25,6 +25,8 @@
   "devDependencies": {
     "@ianvs/prettier-plugin-sort-imports": "^3.7.2",
     "@types/prettier": "^2.7.2",
+    "@types/react": "^18.0.32",
+    "@types/styled-components": "^5.1.26",
     "babel-plugin-styled-components": "^2.0.7",
     "eslint": "8.37.0",
     "eslint-config-next": "13.2.4",
